fix small problems
